VariantVcfFactory and derived classes moved to 'pipeline.io.mappers' package
